# kisaan-backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: KisaanCenter API
  version: 1.0.0
  description: Full API documentation for KisaanCenter backend
servers:
  - url: http://localhost:3000/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Shop:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        owner_id: { type: integer }
        plan_id: { type: integer, nullable: true }
        address: { type: string, nullable: true }
        contact: { type: string, nullable: true }
        status: { type: string, enum: [active, inactive], default: active }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [name, owner_id, address, contact]
    ShopCreate:
      type: object
      properties:
        name: { type: string }
        owner_id: { type: integer }
        plan_id: { type: integer, nullable: true }
        address: { type: string }
        contact: { type: string }
        status: { type: string, enum: [active, inactive], default: active }
      required: [name, owner_id, address, contact]
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category_id: { type: integer }
        status: { type: string, enum: [active, inactive], default: active }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [name, category_id]
...
    Transaction:
      type: object
      properties:
        id: { type: integer }
        shop_id: { type: integer }
        farmer_id: { type: integer }
        buyer_id: { type: integer }
        category_id: { type: integer }
        product_name: { type: string }
        quantity: { type: number, minimum: 0 }
        unit_price: { type: number, minimum: 0 }
        total_sale_value: { type: number, minimum: 0 }
        shop_commission: { type: number, minimum: 0 }
        farmer_earning: { type: number, minimum: 0 }
        status: { type: string, enum: [pending, completed, cancelled], default: pending }
        transaction_date: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [shop_id, farmer_id, buyer_id, category_id, product_name, quantity, unit_price]
...
    ShopCategory:
      type: object
      properties:
        id: { type: integer }
        shop_id: { type: integer }
        category_id: { type: integer }
        is_active: { type: boolean, default: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [shop_id, category_id]

    # Response schemas
    PaginatedResponse:
      type: object
      properties:
        data: { type: array }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, default: false }
        message: { type: string }
        error: { type: string }
...
paths:
...
  /shops:
    post:
      summary: Create shop
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShopCreate' }
      responses:
        '201':
          description: Shop created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Shop' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: Get all shops
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive] }
        - in: query
          name: owner_id
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: List of shops
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Shop' }
  /shops/{id}:
    get:
      summary: Get shop by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Shop found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Shop' }
        '404':
          description: Shop not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      summary: Update shop
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShopCreate' }
      responses:
        '200':
          description: Shop updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Shop' }
        '404':
          description: Shop not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      summary: Delete shop
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Shop deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '404':
          description: Shop not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /transactions/shop/{shopId}:
    get:
      summary: Get transactions by shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Shop transactions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Transaction' }
...
  /balance/shop/{id}:
    get:
      summary: Get shop balance
      parameters:
        - in: path
          name: id
          required: true
...
  /shops/{id}/products:
    get:
      summary: Get products for shop
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Shop products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Product' }
  /shops/{shopId}/products/{productId}:
    post:
      summary: Assign product to shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '201':
          description: Product assigned to shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      summary: Remove product from shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Product removed from shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    patch:
      summary: Toggle product active status for shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Product status toggled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /shop-categories/shop/{shopId}:
    get:
      summary: Get categories for shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
        - in: query
          name: is_active
          schema: { type: boolean }
      responses:
        '200':
          description: Shop categories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ShopCategory' }